# sez/src/main/resources/api/swagger_full.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "API for managing the tutoring system (CRM)",
    "title" : "SEZ Tutoring Platform API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local Server",
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "authenticateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Successful login"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Invalid credentials"
          }
        },
        "summary" : "Authentication via email and password",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/google" : {
      "post" : {
        "operationId" : "authenticateGoogle",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GoogleLoginDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Successful Google authentication"
          }
        },
        "summary" : "Authentication via Google ID Token",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/complete-registration" : {
      "post" : {
        "operationId" : "completeRegistration",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompleteRegistrationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Registration completed"
          }
        },
        "summary" : "Complete profile registration for OAuth users",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/register/tutor" : {
      "post" : {
        "operationId" : "registerTutor",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TutorRegistrationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            },
            "description" : "Tutor registered"
          }
        },
        "summary" : "Register a new tutor",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/register/student" : {
      "post" : {
        "operationId" : "registerStudent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StudentRegistrationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            },
            "description" : "Student registered"
          }
        },
        "summary" : "Register a new student",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/forgot-password" : {
      "post" : {
        "operationId" : "forgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForgotPasswordDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            },
            "description" : "Reset email sent"
          }
        },
        "summary" : "Request password reset link",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/reset-password" : {
      "post" : {
        "operationId" : "resetPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetPasswordDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            },
            "description" : "Password changed successfully"
          },
          "400" : {
            "description" : "Invalid or expired token"
          }
        },
        "summary" : "Set new password using token",
        "tags" : [ "Auth" ]
      }
    },
...
    "/api/v1/assignments/my" : {
      "get" : {
        "operationId" : "getStudentAssignments",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "schema" : {
            "$ref" : "#/components/schemas/AssignmentStatus"
          }
        }, {
          "in" : "query",
          "name" : "courseName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "tutorName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Sort by date. True for newest first, False for oldest first.",
          "in" : "query",
          "name" : "dateDescending",
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AssignmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of student's assignments"
          }
        },
        "summary" : "Get list of assignments for the current student",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/tutor/my" : {
      "get" : {
        "operationId" : "getTutorAssignments",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "schema" : {
            "$ref" : "#/components/schemas/AssignmentStatus"
          }
        }, {
          "in" : "query",
          "name" : "courseId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "studentName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Sort by date. True for newest first, False for oldest first.",
          "in" : "query",
          "name" : "dateDescending",
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AssignmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of assignments for the tutor"
          }
        },
        "summary" : "Get list of assignments managed by the tutor",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/lesson/{lessonId}" : {
      "post" : {
        "operationId" : "createAssignment",
        "parameters" : [ {
          "in" : "path",
          "name" : "lessonId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAssignmentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Assignment created successfully"
          }
        },
        "summary" : "Create a new assignment for a specific lesson",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/{id}" : {
      "patch" : {
        "operationId" : "updateAssignmentInfo",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateAssignmentInfoDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Assignment info updated successfully"
          }
        },
        "summary" : "Update assignment title and description",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/{id}/submit" : {
      "post" : {
        "operationId" : "completeAssignment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Assignment marked as completed"
          }
        },
        "summary" : "Mark assignment as completed by student",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/{id}/grade" : {
      "patch" : {
        "operationId" : "gradeAssignment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GradeAssignmentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Assignment graded successfully"
          }
        },
        "summary" : "Grade the completed assignment",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/{id}/attachments" : {
      "get" : {
        "operationId" : "getAssignmentByLesson",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description" : "Assignment details"
          },
          "404" : {
            "description" : "Assignment not found for this lesson"
          }
        },
        "summary" : "Get assignment details for a specific lesson",
        "tags" : [ "Assignments" ]
      },
      "post" : {
        "operationId" : "uploadAttachment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/uploadAttachment_request"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "File uploaded successfully"
          }
        },
        "summary" : "Upload a file to assignment",
        "tags" : [ "Assignments" ]
      }
    },
    "/api/v1/assignments/attachments/{attachmentId}" : {
      "get" : {
        "operationId" : "downloadAttachment",
        "parameters" : [ {
          "in" : "path",
          "name" : "attachmentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "The file content"
          }
        },
        "summary" : "Download attachment content",
        "tags" : [ "Assignments" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AuthResponse" : {
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "registrationComplete" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "LoginDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
      "GoogleLoginDto" : {
        "properties" : {
          "idToken" : {
            "type" : "string"
          }
        },
        "required" : [ "idToken" ],
        "type" : "object"
      },
      "CompleteRegistrationDto" : {
        "properties" : {
          "role" : {
            "enum" : [ "STUDENT", "TUTOR" ],
            "type" : "string"
          },
          "name" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "surname" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "bio" : {
            "maxLength" : 1000,
            "type" : "string"
          }
        },
        "required" : [ "name", "role", "surname" ],
        "type" : "object"
      },
      "MessageResponse" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TutorRegistrationDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "password" : {
            "maxLength" : 255,
            "minLength" : 8,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "surname" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "bio" : {
            "maxLength" : 1000,
            "type" : "string"
          }
        },
        "required" : [ "email", "name", "password", "surname" ],
        "type" : "object"
      },
      "StudentRegistrationDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "password" : {
            "maxLength" : 255,
            "minLength" : 8,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "surname" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          }
        },
        "required" : [ "email", "name", "password", "surname" ],
        "type" : "object"
      },
      "ForgotPasswordDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "ResetPasswordDto" : {
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 255,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "required" : [ "newPassword", "token" ],
        "type" : "object"
      },
...
      "AssignmentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lessonId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lessonStartTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "tutorFullName" : {
            "type" : "string"
          },
          "studentFullName" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/AssignmentStatus"
          },
          "grade" : {
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer"
          },
          "attachments" : {
            "items" : {
              "$ref" : "#/components/schemas/AttachmentResponse"
            },
            "type" : "array"
          }
        },
        "required" : [ "description", "id", "lessonId", "lessonStartTime", "status", "studentFullName", "title", "tutorFullName" ],
        "type" : "object"
      },
      "AssignmentStatus" : {
        "enum" : [ "PENDING", "COMPLETED", "GRADED" ],
        "type" : "string"
      },
      "AttachmentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreateAssignmentDto" : {
        "properties" : {
          "title" : {
            "maxLength" : 200,
            "minLength" : 3,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 2000,
            "type" : "string"
          }
        },
        "required" : [ "description", "title" ],
        "type" : "object"
      },
      "UpdateAssignmentInfoDto" : {
        "properties" : {
          "title" : {
            "maxLength" : 200,
            "minLength" : 3,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 2000,
            "minLength" : 1,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GradeAssignmentDto" : {
        "properties" : {
          "grade" : {
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer"
          }
        },
        "required" : [ "grade" ],
        "type" : "object"
      },
...
      "uploadAttachment_request" : {
        "properties" : {
          "file" : {
            "format" : "binary",
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  }
}